LibZFS lifetime reduction
Reduce the lifetime of the LibZFS states kept around in ZetaWatch and
ZetaAuthorizationHelper. In the helper, the library is created for each
individual request. In the gui, it is recreated every time the menu is
opened.
This was required because changed properties such as the altroot of
freshly imported pools was not properly picked up when they were
imported, if they were already seen previously. This lead to datasets
of pools being imported with altroot being mounted in an incorrect
mount point.